Moved check and read_model to emit_ir.

+++ tools/mo/openvino/tools/mo/convert_impl.py
@@ -400,6 +400,14 @@ def prepare_ir(argv: argparse.Namespace):
 
 
 def emit_ir(graph: Graph, argv: argparse.Namespace):
+    # We have to separate fe object lifetime from fem to
+    # avoid segfault during object destruction. So fe must
+    # be destructed before fem object explicitly.
+    def read_model(path_to_xml):
+        fe = fem.load_by_framework(framework="ir")
+        function = fe.convert(fe.load(path_to_xml))
+        return function
+
     NormalizeTI().find_and_replace_pattern(graph)
     for_graph_and_each_sub_graph_recursively(graph, RemoveConstOps().find_and_replace_pattern)
     for_graph_and_each_sub_graph_recursively(graph, CreateConstNodesReplacement().find_and_replace_pattern)
@@ -425,37 +433,40 @@ def emit_ir(graph: Graph, argv: argparse.Namespace):
     output_dir = argv.output_dir if argv.output_dir != '.' else os.getcwd()
     orig_model_name = os.path.normpath(os.path.join(output_dir, argv.model_name))
 
+    fem = FrontEndManager()
+    func = read_model(orig_model_name + "_tmp.xml")
+
     return_code = "not executed"
-    func = None
-    try:
-        from openvino.tools.mo.back.offline_transformations import apply_offline_transformations
-        func = apply_offline_transformations(orig_model_name, argv)
-        if "compress_fp16" in argv and argv.compress_fp16:
-            # restore data_type cmd parameter
-            argv.data_type = 'FP16'
-        return_code = 0
-    except Exception as e:
-        return_code = "failed"
-        log.error(e)
+    if not(argv.framework == 'tf' and argv.tensorflow_custom_operations_config_update):
+        try:
+            from openvino.tools.mo.back.offline_transformations import apply_offline_transformations
+            func = apply_offline_transformations(func, argv)
+            if "compress_fp16" in argv and argv.compress_fp16:
+                # restore data_type cmd parameter
+                argv.data_type = 'FP16'
+            return_code = 0
+        except Exception as e:
+            return_code = "failed"
+            log.error(e)
 
-    message = str(dict({
-        "platform": platform.system(),
-        "mo_version": get_simplified_mo_version(),
-        "ie_version": get_simplified_ie_version(env=os.environ),
-        "python_version": sys.version,
-        "return_code": return_code
-    }))
-    t = tm.Telemetry()
-    t.send_event('mo', 'offline_transformations_status', message)
+        message = str(dict({
+            "platform": platform.system(),
+            "mo_version": get_simplified_mo_version(),
+            "ie_version": get_simplified_ie_version(env=os.environ),
+            "python_version": sys.version,
+            "return_code": return_code
+        }))
+        t = tm.Telemetry()
+        t.send_event('mo', 'offline_transformations_status', message)
 
-    if return_code != 0:
-        raise Error("offline transformations step has failed.")
+        if return_code != 0:
+            raise Error("offline transformations step has failed.")
 
-    for suf in [".xml", ".bin", ".mapping"]:
-        # remove existing files
-        path_to_file = orig_model_name + "_tmp" + suf
-        if os.path.exists(path_to_file):
-            os.remove(path_to_file)
+        for suf in [".xml", ".bin", ".mapping"]:
+            # remove existing files
+            path_to_file = orig_model_name + "_tmp" + suf
+            if os.path.exists(path_to_file):
+                os.remove(path_to_file)
     return func
